Add cubemap support & add ktx basic writer implementation
- Fix reading cubemaps, as it was currently unsupported.
- Closes #3 as it was not an appropriate fix for that

I do not believe this still actually properly reads texture
arrays/multi-face textures, but this does handle the common cases of
"standard" textures (which you already implemented) and adds appropriate
support for 6-face textures (cubemaps) for reading and writing.

This PR added an API change to the `texture` iterator for slice, so I
have incremented the version to 0.4 as it is a breaking API change.